requirements.txt
# Core scientific computing
numpy>=2.0.0
scipy>=1.15.0
pandas>=2.0.0

# Audio processing
soundfile>=0.13.0
sounddevice>=0.5.0

# Visualization
matplotlib>=3.10.0

# AWS and cloud storage
boto3>=1.40.0
botocore>=1.40.0
cloudpathlib>=0.23.0

# Data serialization
msgpack>=1.1.0

# Image processing (for MP4 generation scripts)
Pillow>=10.0.0
